feat(device table): move logic according fsd
