+++ src/components/pages/devices/Devices.tsx
@@ -1,13 +1,13 @@
 import { useEffect } from 'react';
 
-import OfficeFileBtn from '@/components/ui/buttons/download/OfficeFileBtn';
 import Loader from '@/components/ui/loader/Loader';
-import DevicesTable from '@/features/devices/ui/DevicesTable';
+import OfficeFileBtn from '@/features/export-table-data/ui/OfficeFileBtn';
 import { setAssignedDevice, setIssueStep } from '@/features/issue/model/issueSlice';
 import { useDeviceTableController } from '@/hooks/data/useDeviceTableController';
 import { useAppDispatch } from '@/hooks/redux/useRedux';
 import { resetDevice, resetStatus } from '@/store/slices/deviceSlice';
 import { FilteredDevicesFromBack } from '@/types/devices';
+import { DevicesTable } from '@/widgets/devices-table/ui/DevicesTable';
 
 import styles from './Devices.module.scss';
