--- src/components/ui/buttons/download/OfficeFileBtn.module.scss
@@ -1,22 +1,0 @@
-.icons {
-    font-size: 18px;
-    .items {
-        display: flex;
-        gap: 6px;
-        cursor: pointer;
-        .green {
-            color: #52627d;
-            transition: color .3s ease-in-out;
-            &:hover {
-                color: #3faaa8;
-            }
-        }
-        .yellow {
-            color: #52627d;
-            transition: color .3s ease-in-out;
-            &:hover {
-                color: #ffbc4b;
-            }
-        }
-    }
-}
