--- src/shared/ui/filterDropdown/styles.module.scss
@@ -1,8 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  padding: 12px 8px;
-  .checkboxes {
-    padding: 12px;
-  }
-}
